backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pycryptodome==3.20.0
pilk==0.2.4
python-multipart==0.0.9
python-jose==3.3.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1
pydantic-settings==2.3.1
fastapi-pagination==0.12.24
cachetools==5.3.3
passlib==1.7.4
bcrypt==3.2.0
pytest==8.2.2
protobuf==5.28.0-rc2
lz4==4.3.3
xmltodict==0.13.0
APScheduler==3.10.1
pyotp==2.9.0
cryptography==44.0.1
zstandard==0.23.0